DBZ-4393 Rename package and modules as schemagenerator
