--- debezium-api-generator/src/main/resources/META-INF/services/io.debezium.api_generator.formats.ApiFormat
@@ -1,1 +1,0 @@
-io.debezium.api_generator.formats.OpenApiFormat
