+++ debezium-bom/pom.xml
@@ -609,7 +609,7 @@
             </dependency>
             <dependency>
                 <groupId>io.debezium</groupId>
-                <artifactId>debezium-api-generator</artifactId>
+                <artifactId>debezium-schema-generator</artifactId>
                 <version>${project.version}</version>
             </dependency>
