+++ debezium-schema-generator/src/main/java/io/debezium/schemagenerator/formats/ApiFormatName.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.api_generator.formats;
+package io.debezium.schemagenerator.formats;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
