+++ debezium-schema-generator/src/main/java/io/debezium/schemagenerator/formats/OpenApiFormat.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.api_generator.formats;
+package io.debezium.schemagenerator.formats;
 
 import java.io.IOException;
 import java.util.Map;
@@ -79,7 +79,7 @@ public class OpenApiFormat implements ApiFormat {
         debeziumAPI.getInfo().setTitle("Generated by Debezium OpenAPI Generator");
 
         debeziumAPI.getInfo().setVersion(
-                IoUtil.loadProperties(OpenApiFormat.class, "io/debezium/api_generator/build.properties").getProperty("version"));
+                IoUtil.loadProperties(OpenApiFormat.class, "io/debezium/schemagenerator/build.properties").getProperty("version"));
 
         debeziumConnectorTypeComponents.addSchema(
                 "debezium-" + connectorSchema.getExtensions().get("connector-id") + "-" + connectorSchema.getExtensions().get("version"),
