+++ debezium-schema-generator/src/main/resources/META-INF/services/io.debezium.schemagenerator.formats.ApiFormat
@@ -1,0 +1,1 @@
+io.debezium.schemagenerator.formats.OpenApiFormat
