+++ debezium-schema-generator/src/main/resources/io/debezium/schemagenerator/build.properties
