+++ pom.xml
@@ -153,7 +153,7 @@
         <module>debezium-server</module>
         <module>debezium-testing</module>
         <module>debezium-connect-rest-extension</module>
-        <module>debezium-api-generator</module>
+        <module>debezium-schema-generator</module>
     </modules>
 
     <distributionManagement>
